requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1
django-colorfield==0.7.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==3.0.1
django-jazzmin==2.5.0
et-xmlfile==1.1.0
fontawesomefree==6.2.0
future==0.18.2
html5lib==1.1
idna==3.4
lxml==4.9.1
MarkupPy==1.14
MarkupSafe==2.1.1
matplotlib==3.6.2
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
Pillow==9.2.0
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.6
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
svglib==1.4.1
tablib==3.2.1
tinycss2==1.2.1
tqdm==4.64.1
tzdata==2022.2
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.13
webencodings==0.5.1
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
